\name{fastConcProb-package}
\alias{fastConcProb-package}
\docType{package}
\title{
  Fast exact calculation of concordance probability
}
\description{
  This package gives two exact methods to calculate the concordance probability in an efficient way.
}
\details{
  \tabular{ll}{
    Package: \tab fastConcProb\cr
    Type: \tab Package\cr
    Version: \tab 1.0.0\cr
    License: \tab GPL (>= 2)\cr
  }
}

\author{
  Ponnet J., Raymaekers J., Van Oirbeek R., Verdonck T.\cr
  Maintainer: Ponnet J. <jolien.ponnet@kuleuven.be>
}

